+++ docs/pages/Layout.vue
@@ -2,10 +2,10 @@
     <div class="content">
         <p class="subtitle">Layout</p>
         <ul>
-            <li><a href="http://bulma.io/documentation/layout/container/" target="_blank">Container</a></li>
-            <li><a href="http://bulma.io/documentation/layout/hero/" target="_blank">Hero</a></li>
-            <li><a href="http://bulma.io/documentation/layout/section/" target="_blank">Section</a></li>
-            <li><a href="http://bulma.io/documentation/layout/footer/" target="_blank">Footer</a></li>
+            <li><a href="https://bulma.io/documentation/layout/container/" target="_blank">Container</a></li>
+            <li><a href="https://bulma.io/documentation/layout/hero/" target="_blank">Hero</a></li>
+            <li><a href="https://bulma.io/documentation/layout/section/" target="_blank">Section</a></li>
+            <li><a href="https://bulma.io/documentation/layout/footer/" target="_blank">Footer</a></li>
             <li><a href="https://bulma.io/documentation/layout/tiles/" target="_blank">Tile system</a></li>
             <li><a href="https://bulma.io/documentation/layout/level/" target="_blank">Level</a></li>
             <li><a href="https://bulma.io/documentation/layout/media-object/" target="_blank">Media</a></li>
@@ -20,9 +20,9 @@
 
         <p class="subtitle">Elements</p>
         <ul>
-            <li><a href="http://bulma.io/documentation/elements/button/" target="_blank">Button</a></li>
-            <li><a href="http://bulma.io/documentation/components/navbar/" target="_blank">Navbar</a></li>
-            <li><a href="http://bulma.io/documentation/components/card/" target="_blank">Card</a></li>
+            <li><a href="https://bulma.io/documentation/elements/button/" target="_blank">Button</a></li>
+            <li><a href="https://bulma.io/documentation/components/navbar/" target="_blank">Navbar</a></li>
+            <li><a href="https://bulma.io/documentation/components/card/" target="_blank">Card</a></li>
         </ul>
 
         <p class="subtitle">